fix(ui): pool reply-less wake activity into one chat rollup
Consecutive runs whose entire visible outcome is tool activity (heartbeat
wakes, cron ticks) now collapse into a single expandable activity row
instead of stacking identical rows down the transcript. Runs with a
visible reply keep per-run activity separation. Group summaries reuse the
standalone row's tool display label, so the rollup reads
"Used Heartbeat Respond ×N". Splits tool call/result pairing out of
chat-thread-grouping.ts into chat-tool-activity-coalesce.ts (max-lines).
